+++ keeper/keeper_test.go
@@ -9,17 +9,16 @@ import (
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 
-	ibctesting "github.com/bianjieai/nft-transfer/testing"
-	ics721testing "github.com/bianjieai/nft-transfer/testing"
 	"github.com/bianjieai/nft-transfer/testing/mock"
 	"github.com/bianjieai/nft-transfer/testing/simapp"
 	"github.com/bianjieai/nft-transfer/types"
+	ibctesting "github.com/cosmos/ibc-go/v10/testing"
 )
 
 type KeeperTestSuite struct {
 	suite.Suite
 
-	coordinator *ics721testing.Coordinator
+	coordinator *ibctesting.Coordinator
 
 	// testing chains used for convenience and readability
 	chainA *ibctesting.TestChain
@@ -33,7 +32,7 @@ type KeeperTestSuite struct {
 }
 
 func (suite *KeeperTestSuite) SetupTest() {
-	suite.coordinator = ics721testing.NewCoordinator(suite.T(), 3)
+	suite.coordinator = ibctesting.NewCustomAppCoordinator(suite.T(), 3, simapp.SetupTestingApp)
 	suite.chainA = suite.coordinator.GetChain(ibctesting.GetChainID(1))
 	suite.chainB = suite.coordinator.GetChain(ibctesting.GetChainID(2))
 	suite.chainC = suite.coordinator.GetChain(ibctesting.GetChainID(3))